web/requirements.txt
aiohttp
python-socketio
setproctitle
numpy
orjson==3.8.2
